+++ lprint-zj.h
@@ -1,0 +1,14 @@
+//
+// ZIJIANG driver table for LPrint, a Label Printer Application
+//
+// Copyright © 2022 by Abdelhakim Qbaich.
+// Copyright © 2019-2022 by Michael R Sweet.
+//
+// Licensed under Apache License v2.0.  See the file "LICENSE" for more
+// information.
+//
+
+{ "zj_58mm",        "Zijiang 58mm Receipt Printer",
+  "MFG:Zijiang;MDL:ZJ-58;", NULL },
+{ "zj_80mm",        "Zijiang 80mm Receipt Printer",
+  "MFG:Zijiang;MDL:ZJ-80;", NULL },