+++ lprint.h
@@ -64,6 +64,8 @@ extern bool	lprintDYMO(pappl_system_t *system, const char *driver_name, const ch
 extern bool	lprintEPL2(pappl_system_t *system, const char *driver_name, const char *device_uri, const char *device_id, pappl_pr_driver_data_t *data, ipp_t **attrs, void *cbdata);
 extern bool	lprintTestFilterCB(pappl_job_t *job, pappl_device_t *device, void *data);
 extern const char *lprintTestPageCB(pappl_printer_t *printer, char *buffer, size_t bufsize);
+extern bool	lprintZJ(pappl_system_t *system, const char *driver_name, const char *device_uri, const char *device_id, pappl_pr_driver_data_t *data, ipp_t **attrs, void *cbdata);
+extern void	lprintZJQueryDriver(pappl_system_t *system, const char *device_uri, char *name, size_t namesize);
 extern bool	lprintZPL(pappl_system_t *system, const char *driver_name, const char *device_uri, const char *device_id, pappl_pr_driver_data_t *data, ipp_t **attrs, void *cbdata);
 extern void	lprintZPLQueryDriver(pappl_system_t *system, const char *device_uri, char *name, size_t namesize);